[change] Icon 컴포넌트 extends로 구조 변경 (AnimationIcon, EtcIcon, Icon)
