+++ src/Elements/Core/Icon/EtcIcon.vue
@@ -1,0 +1,27 @@
+<script>
+import IconFrame from '@/src/Elements/Core/Icon/IconFrame';
+import IconDropdownMediumLineEtc from '@/assets/images/icon/etc/icon-dropdown-medium-line-etc.svg?inline';
+import IconDropdownSmallFillEtc from '@/assets/images/icon/etc/icon-dropdown-small-fill-etc.svg?inline';
+import customValidator from '@/utils/custom-validator';
+
+export const EtcIconNames = ['IconDropdownMediumLineEtc', 'IconDropdownSmallFillEtc'];
+
+export default {
+	name: 'EtcIcon',
+	extends: IconFrame,
+	props: {
+		name: {
+			type: String,
+			default: null,
+			validator(value) {
+				const isValid = EtcIconNames.indexOf(value) !== -1;
+				return customValidator(value, isValid, 'EtcIcon', 'name');
+			},
+		},
+	},
+	components: {
+		IconDropdownMediumLineEtc,
+		IconDropdownSmallFillEtc,
+	},
+};
+</script>
