+++ src/Elements/Core/Icon/index.js
@@ -1,4 +1,5 @@
 import AnimationIcon from './AnimationIcon';
+import EtcIcon from './EtcIcon';
 import Icon from './Icon';
 
-export { AnimationIcon, Icon };
+export { AnimationIcon, EtcIcon, Icon };
